requirements.txt
discord.py==2.3.2
requests==2.31.0
PyYAML==6.0.1
python-dotenv==1.0.0
colorama==0.4.6
humanize==4.8.0
asyncio